fix enum

+++ src/kudu/tools/ksck-test.cc
@@ -389,7 +389,7 @@ class KsckTest : public KuduTest {
   const string KsckResultsToJsonString() {
     ostringstream json_stream;
     ksck_->results().PrintJsonTo(PrintMode::JSON_COMPACT,
-                                 (int)PrintSections::DEFAULT_PRINT_SECTIONS, json_stream);
+                                 PrintSections::DEFAULT_PRINT_SECTIONS, json_stream);
     return json_stream.str();
   }
